Move shared logic into trait
