+++ src/BC/Crypto/Crypto.php
@@ -44,9 +44,12 @@ use Throwable;
  */
 final readonly class Crypto implements CryptoInterface
 {
+    use ContextualEncryptionTrait;
+
     public function __construct(
         private KeychainInterface $keychain,
         private LoggerInterface $logger,
+        private bool $shouldEncryptForDatabase,
         private bool $shouldEncryptForFilesystem,
     ) {
     }
@@ -133,54 +136,5 @@ final readonly class Crypto implements CryptoInterface
         } catch (Throwable) {
             return false;
         }
-    }
-
-    public function encryptForDatabase(?string $value): string
-    {
-        if ($value === null || $value === '') {
-            return '';
-        }
-        return $this->encryptStandard($value, keySource: KeySource::Drive);
-    }
-
-    public function encryptForFilesystem(?string $value): string
-    {
-        if ($value === null || $value === '') {
-            return '';
-        }
-        if (!$this->shouldEncryptForFilesystem) {
-            return $value;
-        }
-        return $this->encryptStandard($value, keySource: KeySource::Database);
-    }
-
-    public function decryptFromDatabase(?string $value, ?int $minimumVersion = null): string
-    {
-        if ($value === null || $value === '') {
-            return '';
-        }
-        if (!$this->cryptCheckStandard($value)) {
-            return $value;
-        }
-        $result = $this->decryptStandard($value, keySource: KeySource::Drive, minimumVersion: $minimumVersion);
-        if ($result === false) {
-            throw new CryptoGenException('Decryption failed');
-        }
-        return $result;
-    }
-
-    public function decryptFromFilesystem(?string $value): string
-    {
-        if ($value === null || $value === '') {
-            return '';
-        }
-        if (!$this->cryptCheckStandard($value)) {
-            return $value;
-        }
-        $result = $this->decryptStandard($value, keySource: KeySource::Database);
-        if ($result === false) {
-            throw new CryptoGenException('Decryption failed');
-        }
-        return $result;
     }
 }
